refactored statistics
