+++ mteb/abstasks/AbsTask.py
@@ -17,10 +17,11 @@ from datasets import Dataset, DatasetDict
 from sklearn.preprocessing import MultiLabelBinarizer
 
 from mteb.abstasks.stratification import _iterative_train_test_split
-from mteb.abstasks.TaskMetadata import DescriptiveStatistics, TaskMetadata
+from mteb.abstasks.TaskMetadata import TaskMetadata
 from mteb.encoder_interface import Encoder
 from mteb.languages import LanguageScripts
 from mteb.types import HFSubset, ScoresDict
+from mteb.types.statistics import DescriptiveStatistics
 
 logger = logging.getLogger(__name__)
