+++ mteb/abstasks/AbsTaskBitextMining.py
@@ -5,9 +5,9 @@ from typing import Any
 
 from datasets import Dataset, DatasetDict
 
-from mteb.abstasks.TaskMetadata import DescriptiveStatistics
 from mteb.encoder_interface import Encoder
 from mteb.types import HFSubset, ScoresDict
+from mteb.types.statistics import DescriptiveStatistics
 
 from ..evaluation.evaluators import BitextMiningEvaluator
 from .AbsTask import AbsTask
