+++ mteb/abstasks/AbsTaskSummarization.py
@@ -6,9 +6,9 @@ from typing import Any
 import numpy as np
 from datasets import Dataset
 
-from mteb.abstasks.TaskMetadata import DescriptiveStatistics
 from mteb.encoder_interface import Encoder
 from mteb.types import ScoresDict
+from mteb.types.statistics import DescriptiveStatistics
 
 from ..evaluation.evaluators import SummarizationEvaluator
 from .AbsTask import AbsTask
