+++ mteb/abstasks/Image/AbsTaskAny2AnyMultiChoice.py
@@ -13,8 +13,8 @@ from datasets import Features, Value, load_dataset
 from PIL import Image
 
 from mteb.abstasks.AbsTask import AbsTask
-from mteb.abstasks.TaskMetadata import DescriptiveStatistics
 from mteb.types import ScoresDict
+from mteb.types.statistics import DescriptiveStatistics
 
 from ...evaluation.evaluators import Any2AnyMultiChoiceEvaluator
