+++ mteb/abstasks/Image/AbsTaskImageClustering.py
@@ -6,8 +6,8 @@ from typing import Any
 
 from datasets import Dataset
 
-from mteb.abstasks.TaskMetadata import DescriptiveStatistics
-from mteb.types import ScoresDict, HFSubset
+from mteb.types import HFSubset, ScoresDict
+from mteb.types.statistics import DescriptiveStatistics
 
 from ...encoder_interface import Encoder
 from ...evaluation.evaluators import ImageClusteringEvaluator
