+++ mteb/types/_string_validators.py
