+++ mteb/types/_task_metadata.py
@@ -1,0 +1,1 @@
+from __future__ import annotations
